Refactored code to have themed componenents in seperate folder
