+++ amplifytassel/src/app/components/ForumsNewPost.js
@@ -5,7 +5,7 @@ import MuiAvatar from '@mui/material/Avatar';
 import MuiBox from '@mui/material/Box';
 import MuiPaper from '@mui/material/Paper';
 import TextField from '@mui/material/TextField';
-import ThemedButton from './ThemedButton';
+import ThemedButton from './Themed/ThemedButton';
 import SendRoundedIcon from '@mui/icons-material/SendRounded';
 import useAuth from '../util/AuthContext';
 import { Storage } from 'aws-amplify';
