+++ amplifytassel/src/app/components/NavBarLoggedOut.js
@@ -2,7 +2,7 @@ import React from "react";
 import { Link } from "react-router-dom";
 import Box from "@mui/material/Box";
 import Toolbar from "@mui/material/Toolbar";
-import ThemedButton from "./ThemedButton";
+import ThemedButton from "./Themed/ThemedButton";
 import PersonIcon from "@mui/icons-material/Person";
 import SchoolIcon from "@mui/icons-material/School";
 import * as Nav from "./NavBarComponents";
