+++ amplifytassel/src/app/components/OpportunityForm.js
@@ -24,7 +24,7 @@ import { yupResolver } from "@hookform/resolvers/yup";
 import * as Yup from "yup";
 import { useForm } from "react-hook-form";
 
-import ThemedButton from "../components/ThemedButton";
+import ThemedButton from "./Themed/ThemedButton.js";
 import { TextInput } from "./TextInput";
 import { TimeInput } from "./TimeInput.js";
 import { DropdownInput } from "./DropdownInput";
