+++ amplifytassel/src/app/components/ViewOpportunityAbout.js
@@ -6,7 +6,7 @@ import MuiAccordionSummary from '@mui/material/AccordionSummary';
 import MuiAccordionDetails from '@mui/material/AccordionDetails';
 import MuiPaper from '@mui/material/Paper';
 import ArrowForwardIosSharpIcon from '@mui/icons-material/ArrowForwardIosSharp';
-import ThemedButton from './ThemedButton';
+import ThemedButton from './Themed/ThemedButton';
 import RequestModal from './RequestOpportunityModal';
 import useAuth from '../util/AuthContext';
 import {toast} from 'react-toastify';
