+++ amplifytassel/src/app/components/VolunteerExperienceForm.js
@@ -12,7 +12,7 @@ import * as Yup from 'yup';
 import {TextInput2} from './TextInput2';
 import {DateInput2} from './DateInput2';
 import {CheckboxInput2} from './CheckboxInput2';
-import ThemedButton from '../components/ThemedButton';
+import ThemedButton from './Themed/ThemedButton';
 import useAuth from '../util/AuthContext';
 import {sortWorkExperience} from './WorkExperienceForm';
 import useAnimation from '../util/AnimationContext';
