+++ amplifytassel/src/app/components/WorkExperienceDeleteModal.js
@@ -5,7 +5,7 @@ import Paper from '@mui/material/Paper';
 import {styled} from '@mui/material/styles';
 import {toast} from 'react-toastify';
 
-import ThemedButton from './ThemedButton';
+import ThemedButton from './Themed/ThemedButton';
 import useAuth from '../util/AuthContext';
 import CloseRoundedIcon from '@mui/icons-material/CloseRounded';
 import ButtonBase from '@mui/material/ButtonBase';
