+++ amplifytassel/src/app/pages/UpdateProfile.js
@@ -6,13 +6,13 @@ import MuiBox from "@mui/material/Box";
 import Chip from "@mui/material/Chip";
 import { toast } from "react-toastify";
 import useAuth from "../util/AuthContext";
-import ThemedInput from "../components/ThemedInput";
-import ThemedButton from "../components/ThemedButton";
-import { InputContext } from "../components/ThemedInput";
+import ThemedInput from "../components/Themed/ThemedInput";
+import ThemedButton from "../components/Themed/ThemedButton";
+import { InputContext } from "../components/Themed/ThemedInput";
 import { MultiSelectContext } from "../components/MultiSelect";
 import { Link } from "react-router-dom";
 import WorkExperienceList from "../components/WorkExperienceList";
-import ThemedDropdown from "../components/ThemedDropdown";
+import ThemedDropdown from "../components/Themed/ThemedDropdown";
 import AddIcon from "@mui/icons-material/Add";
 import RemoveIcon from "@mui/icons-material/Remove";
 import ButtonBase from "@mui/material/ButtonBase";
